# contracts/openapi/v1.json
{
  "components": {
    "schemas": {
...
      "ConfidenceIndicator": {
        "description": "Coarse public representation of internal coordinate confidence.",
        "enum": [
          "low",
          "medium",
          "high"
        ],
        "title": "ConfidenceIndicator",
        "type": "string"
      },
      "ContentType": {
        "description": "Public offer granularity.",
        "enum": [
          "development",
          "unit"
        ],
        "title": "ContentType",
        "type": "string"
      },
...
      "LocationMapFeature": {
        "additionalProperties": false,
        "description": "One accepted grouped location feature.",
        "properties": {
          "geometry": {
            "$ref": "#/components/schemas/PointGeometry"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "properties": {
            "$ref": "#/components/schemas/LocationMapProperties"
          },
          "type": {
            "const": "Feature",
            "default": "Feature",
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "id",
          "geometry",
          "properties"
        ],
        "title": "LocationMapFeature",
        "type": "object"
      },
      "LocationMapProperties": {
        "additionalProperties": false,
        "description": "Backend-owned grouped location display data.",
        "properties": {
          "area_max_sqm": {
            "anyOf": [
              {
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Area Max Sqm"
          },
          "area_min_sqm": {
            "anyOf": [
              {
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Area Min Sqm"
          },
          "confidence": {
            "$ref": "#/components/schemas/ConfidenceIndicator"
          },
          "coordinate_precision": {
            "title": "Coordinate Precision",
            "type": "string"
          },
          "currency": {
            "const": "PLN",
            "default": "PLN",
            "title": "Currency",
            "type": "string"
          },
          "display_address": {
            "title": "Display Address",
            "type": "string"
          },
          "display_name": {
            "title": "Display Name",
            "type": "string"
          },
          "district": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "District"
          },
          "latest_published_at": {
            "format": "date-time",
            "title": "Latest Published At",
            "type": "string"
          },
          "matching_offer_count": {
            "minimum": 1.0,
            "title": "Matching Offer Count",
            "type": "integer"
          },
          "price_max_minor": {
            "anyOf": [
              {
                "minimum": 0.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Price Max Minor"
          },
          "price_min_minor": {
            "anyOf": [
              {
                "minimum": 0.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Price Min Minor"
          },
          "total_offer_count": {
            "minimum": 1.0,
            "title": "Total Offer Count",
            "type": "integer"
          }
        },
        "required": [
          "display_name",
          "display_address",
          "district",
          "coordinate_precision",
          "confidence",
          "matching_offer_count",
          "total_offer_count",
          "latest_published_at"
        ],
        "title": "LocationMapProperties",
        "type": "object"
      },
      "LocationMapResponse": {
        "additionalProperties": false,
        "description": "GeoJSON FeatureCollection with stable WEF metadata.",
        "properties": {
          "features": {
            "items": {
              "$ref": "#/components/schemas/LocationMapFeature"
            },
            "title": "Features",
            "type": "array"
          },
          "meta": {
            "$ref": "#/components/schemas/MapResponseMeta"
          },
          "type": {
            "const": "FeatureCollection",
            "default": "FeatureCollection",
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "features",
          "meta"
        ],
        "title": "LocationMapResponse",
        "type": "object"
      },
      "MapResponseMeta": {
        "additionalProperties": false,
        "description": "Bounded response metadata safe for public clients.",
        "properties": {
          "feature_count": {
            "minimum": 0.0,
            "title": "Feature Count",
            "type": "integer"
          },
          "matching_offer_count": {
            "minimum": 0.0,
            "title": "Matching Offer Count",
            "type": "integer"
          },
          "request_id": {
            "format": "uuid",
            "title": "Request Id",
            "type": "string"
          }
        },
        "required": [
          "request_id",
          "feature_count",
          "matching_offer_count"
        ],
        "title": "MapResponseMeta",
        "type": "object"
      },
      "MarketType": {
        "description": "Source market classification.",
        "enum": [
          "primary",
          "secondary",
          "unknown"
        ],
        "title": "MarketType",
        "type": "string"
      },
      "PointGeometry": {
        "additionalProperties": false,
        "description": "GeoJSON Point geometry in longitude/latitude order.",
        "properties": {
          "coordinates": {
            "maxItems": 2,
            "minItems": 2,
            "prefixItems": [
              {
                "type": "number"
              },
              {
                "type": "number"
              }
            ],
            "title": "Coordinates",
            "type": "array"
          },
          "type": {
            "const": "Point",
            "default": "Point",
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "coordinates"
        ],
        "title": "PointGeometry",
        "type": "object"
      },
      "ProblemResponse": {
        "additionalProperties": false,
        "description": "Stable public error envelope without rejected values.",
        "properties": {
          "code": {
            "title": "Code",
            "type": "string"
          },
          "detail": {
            "title": "Detail",
            "type": "string"
          },
          "instance": {
            "title": "Instance",
            "type": "string"
          },
          "kind": {
            "const": "validation_error",
            "default": "validation_error",
            "title": "Kind",
            "type": "string"
          },
          "request_id": {
            "format": "uuid",
            "title": "Request Id",
            "type": "string"
          },
          "status": {
            "title": "Status",
            "type": "integer"
          },
          "title": {
            "title": "Title",
            "type": "string"
          },
          "type": {
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "type",
          "title",
          "status",
          "code",
          "request_id",
          "detail",
          "instance"
        ],
        "title": "ProblemResponse",
        "type": "object"
      }
    }
  },
  "info": {
    "title": "Warsaw Estate Finder API",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/map/locations": {
      "get": {
        "description": "Return grouped accepted locations for the normalized filter query.",
        "operationId": "queryMapLocations",
        "parameters": [
          {
            "in": "query",
            "name": "bbox",
            "required": true,
            "schema": {
              "maxLength": 100,
              "minLength": 7,
              "title": "Bbox",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "price_min",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maximum": 1000000000,
                  "minimum": 0,
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Price Min"
            }
          },
          {
            "in": "query",
            "name": "price_max",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maximum": 1000000000,
                  "minimum": 0,
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Price Max"
            }
          },
          {
            "in": "query",
            "name": "area_min",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "exclusiveMinimum": 0.0,
                  "maximum": 100000.0,
                  "type": "number"
                },
                {
                  "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Area Min"
            }
          },
          {
            "in": "query",
            "name": "area_max",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "exclusiveMinimum": 0.0,
                  "maximum": 100000.0,
                  "type": "number"
                },
                {
                  "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Area Max"
            }
          },
          {
            "in": "query",
            "name": "rooms",
            "required": false,
            "schema": {
              "items": {
                "maximum": 20,
                "minimum": 0,
                "type": "integer"
              },
              "maxItems": 10,
              "title": "Rooms",
              "type": "array"
            }
          },
          {
            "in": "query",
            "name": "district",
            "required": false,
            "schema": {
              "items": {
                "maxLength": 80,
                "minLength": 1,
                "type": "string"
              },
              "maxItems": 20,
              "title": "District",
              "type": "array"
            }
          },
          {
            "in": "query",
            "name": "market_type",
            "required": false,
            "schema": {
              "items": {
                "$ref": "#/components/schemas/MarketType"
              },
              "maxItems": 5,
              "title": "Market Type",
              "type": "array"
            }
          },
          {
            "in": "query",
            "name": "content_type",
            "required": false,
            "schema": {
              "items": {
                "$ref": "#/components/schemas/ContentType"
              },
              "maxItems": 5,
              "title": "Content Type",
              "type": "array"
            }
          },
          {
            "in": "query",
            "name": "published_from",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date-time",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Published From"
            }
          },
          {
            "in": "query",
            "name": "published_to",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date-time",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Published To"
            }
          },
          {
            "in": "header",
            "name": "if-none-match",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "If-None-Match"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LocationMapResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "304": {
            "description": "The filtered projection has not changed."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemResponse"
                }
              }
            },
            "description": "The query is malformed, contradictory, or unsafe."
          }
        },
        "summary": "Query grouped map locations",
        "tags": [
          "map"
        ]
      }
    }
  }
}